Resolved Conflic between Contact-and-Chat-Screens-

+++ app/src/main/java/com/example/smiletalk/ContactsActivity.java
@@ -1,6 +1,7 @@
 package com.example.smiletalk;
 
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatDelegate;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
@@ -11,6 +12,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
+import android.widget.Button;
 
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
@@ -20,7 +22,7 @@ import java.util.List;
 
 
 
-public class ContactsActivity extends AppCompatActivity implements AddContactListener, DeleteContactListener  {
+public class ContactsActivity extends DarkAppCompact implements AddContactListener, DeleteContactListener  {
 
 
     private RecyclerView rvContacts;
@@ -58,6 +60,14 @@ public class ContactsActivity extends AppCompatActivity implements AddContactLis
         // Initialize RecyclerView
         rvContacts = findViewById(R.id.rvContacts);
         rvContacts.setLayoutManager(new LinearLayoutManager(this));
+
+        @Override
+        public void onBackPressed() {
+            // Disable the back button functionality
+            // Remove the following line if you want to enable the back button
+            // super.onBackPressed();
+        }
+
 
         // Prepare data
         curUser = Login.curUser;
@@ -101,14 +111,46 @@ public class ContactsActivity extends AppCompatActivity implements AddContactLis
             }
         });
 
+        Button logoutButton = findViewById(R.id.actionButton);
+
+        logoutButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                showLogoutDialog();
+            }
+        });
+
     }
+
+    public void showLogoutDialog() {
+        Fragment addContactFragment = getSupportFragmentManager().findFragmentByTag("AddContactFragment");
+        Fragment deleteContactFragment = getSupportFragmentManager().findFragmentByTag("DeleteContactFragment");
+        Fragment logoutFragment = getSupportFragmentManager().findFragmentByTag("LogoutFragment");
+
+        if (addContactFragment == null && deleteContactFragment == null && logoutFragment == null) {
+            LogoutFragment fragment = new LogoutFragment();
+            // Pass the index of the item to the fragment
+
+
+            getSupportFragmentManager()
+                    .beginTransaction()
+                    .add(android.R.id.content, fragment, "LogoutFragment")
+                    .addToBackStack(null)
+                    .commit();
+
+            findViewById(R.id.grayOutOverlay).setVisibility(View.VISIBLE);
+        }
+    }
+
+
 
     public void showDeleteContactDialog(int position) {
         // Check if the AddContactFragment is already added
         Fragment addContactFragment = getSupportFragmentManager().findFragmentByTag("AddContactFragment");
         Fragment deleteContactFragment = getSupportFragmentManager().findFragmentByTag("DeleteContactFragment");
+        Fragment logoutFragment = getSupportFragmentManager().findFragmentByTag("LogoutFragment");
 
-        if (addContactFragment == null && deleteContactFragment == null) {
+        if (addContactFragment == null && deleteContactFragment == null && logoutFragment == null) {
             // Neither AddContactFragment nor DeleteContactFragment is added, proceed with adding DeleteContactFragment
             DeleteContactFragment fragment = new DeleteContactFragment(position);
             // Pass the index of the item to the fragment
@@ -163,7 +205,7 @@ public class ContactsActivity extends AppCompatActivity implements AddContactLis
         // Hide the overlay
         findViewById(R.id.grayOutOverlay).setVisibility(View.GONE);
     }
-    @Override
+        @Override
     protected void onResume() {
         super.onResume();
         // Set the updated adapter to the RecyclerView
